pyproject.toml
[project]
name = "pygame-animated-sprite"
version = "2.0.0-alpha"
authors = [
    {name = "Pio", email = "pioooop2@gmail.com"},
]
description = ""
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pygame-ce"
]
